spllit index content
