+++ Read Model/ReadModel.Domain/Index/HttpRequest/HttpRequest.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Nest;
 
 namespace ReadModel.Domain.Index.HttpRequest
 {
@@ -16,36 +15,5 @@ namespace ReadModel.Domain.Index.HttpRequest
         public string FormattedDate { get; set; }
         public TimeSpan Duration { get; set; }
         public string FormattedDuration { get; set; }
-    }
-
-    public class GeoIp
-    {
-        public string IPv4 { get; set; }
-        public GeoLocation Location { get; set; }
-    }
-
-    public class HttpRequestRequest
-    {
-        public string Method { get; set; }
-        public string Uri { get; set; }
-        public string Headers { get; set; }
-        public string Body { get; set; }
-        public string Os { get; set; }
-        public string Browser { get; set; }
-        public string Device { get; set; }
-    }
-
-    public class HttpRequestResponse
-    {
-        public string Body { get; set; }
-        public int HttpStatus { get; set; }
-    }
-
-    public class HttpRequestAudit
-    {
-        public string ClientApplication { get; set; }
-        public Guid CorrelationId { get; set; }
-        public Guid UserId { get; set; }
-        public Guid ImpersonatedUserId { get; set; }
     }
 }
