+++ Read Model/ReadModel.Domain/Index/HttpRequest/HttpRequestResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace ReadModel.Domain.Index.HttpRequest
+{
+    public class HttpRequestResponse
+    {
+        public string Body { get; set; }
+        public int HttpStatus { get; set; }
+    }
+}
